Gesichtserkennung/requirements.txt
click==8.1.8
colorama==0.4.6
dlib
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy
opencv-python==4.11.0.86
pillow==11.1.0
websocket-client==1.8.0
